feat(front): #373 add imgs for datadose
- add imgs in data for datadose
- add new type for imgs in datadose type
- refact data only for tips
- add new file in utils for datadose elements
