+++ packages/front/src/components/DataDose/types.ts
@@ -1,7 +1,12 @@
+import type { ImageSourcePropType } from "react-native";
+
+import type { imgs } from "assets/imgs";
 import type { TDiseases } from "utils/db/types";
 
+type ImgName = keyof typeof imgs;
 interface TDataDose {
   details: string;
+  img: ImageSourcePropType | ImgName;
   tip: string;
 }
